move state to common
