+++ core/src/commands/follow_chain.rs
@@ -28,11 +28,11 @@ use sp_runtime::{
 use substrate_rpc_client::{ws_client, ChainApi, FinalizedHeaders, Subscription, WsClient};
 
 use crate::{
-    build_executor,
-    common::parse,
-    full_extensions, rpc_err_handler,
-    state::{LiveState, RuntimeChecks, State},
-    state_machine_call_with_proof, SharedParams, LOG_TARGET,
+    common::{
+        parse,
+        state::{build_executor, state_machine_call_with_proof, LiveState, RuntimeChecks, State},
+    },
+    full_extensions, rpc_err_handler, SharedParams, LOG_TARGET,
 };
 
 const SUB: &str = "chain_subscribeFinalizedHeads";
