+++ core/src/commands/offchain_worker.rs
@@ -23,11 +23,11 @@ use sp_runtime::traits::{Block as BlockT, NumberFor};
 use substrate_rpc_client::{ws_client, ChainApi};
 
 use crate::{
-    build_executor,
-    common::parse,
-    full_extensions, rpc_err_handler,
-    state::{LiveState, RuntimeChecks, State},
-    state_machine_call, SharedParams, LOG_TARGET,
+    common::{
+        parse,
+        state::{build_executor, state_machine_call, LiveState, RuntimeChecks, State},
+    },
+    full_extensions, rpc_err_handler, SharedParams, LOG_TARGET,
 };
 
 /// Configuration for [`run`].
